fix: ensure legacy hook types are populated
